Started restructuration
